refactor: move session-token logic to shared/logic
