+++ apps/user/package.json
@@ -16,9 +16,6 @@
   },
   "dependencies": {
     "@inlang/paraglide-sveltekit": "^0.12.1",
-    "@node-rs/argon2": "^2.0.2",
-    "@oslojs/crypto": "^1.0.1",
-    "@oslojs/encoding": "^1.1.0",
     "@paralleldrive/cuid2": "^2.2.2",
     "drizzle-orm": "^0.37.0",
     "std-env": "^3.8.0",
